refactor: organismsディレクトリ内のゲームボード関連コンポーネントをboard/サブディレクトリに整理
- DuelField, ExtraDeck, Graveyard, Hands, MainDeckをboard/サブディレクトリに移動
- DuelField.svelteとシミュレーターページのimportパスを更新
- コンポーネントの責務を明確化し、ディレクトリ構造を改善
