+++ skeleton-app/src/lib/components/organisms/board/Graveyard.svelte
