+++ skeleton-app/src/lib/components/organisms/board/Hands.svelte
